perf: keep the secp256k1 build out of the initial bundle of most routes
WalletService binds ecpair/tiny-secp256k1 (the ~1.7 MB asm.js secp256k1
build) at module load, and it was statically reachable from the root
layout's providers — so every route, even /terms and /about, shipped the
crypto in its initial bundle.

The eager path is now free of it:

- A new lean encryption.ts holds the scrypt + AES-GCM + CryptoJS helpers,
  which have no elliptic-curve dependency. StorageService, BackupService
  and SecurityService — all instantiated eagerly — import from it instead
  of from WalletService. StorageService keeps only a type-only import of
  AddressType, which is erased.
- SecurityService loads ECC lazily (it already had initializeCrypto() for
  exactly this) for the legacy-upgrade WIF check, and bip39 lazily at its
  one use site.
- ElectrumService loads ecpair/tiny-secp256k1 lazily inside
  getPublicKeyForAddress, its only consumer; the read-only balance path
  no longer pulls them.
- WalletContext and AuthenticationDialog import WalletService dynamically,
  so it loads after first paint rather than blocking it.
- Explorer-URL helpers move to a crypto-free lib/explorer, freeing
  TransactionHistory, and the home page loads SendForm via next/dynamic so
  the send crypto loads only when the Send tab is opened.

First Load JS, before -> after:

  /terms, /about, /settings, most /settings/*   ~1.86 MB -> ~430 kB  (-77%)
  /  (balance + history render immediately)      1.94 MB -> 455 kB   (-77%)

Routes whose primary job is a key operation — /onboarding, /settings/wallet,
/settings/advanced, /connect — still load crypto up front, since deferring
it there would render a form that cannot work yet. Splitting those is a
follow-up.

Verified: 542 unit tests pass, typecheck clean, zero lint errors, and the
end-to-end suite passes — which exercises the changed import timing on the
unlock, wallet-init and sign flows.
